Added image rotate and made rotate90/180/270 internal
No longer copy source SDL_Surfaces if they're already in the desired RBGA format.
The general "rotate" uses the 3 sheer method to rotate +/- 45 degrees from one of the 4 (0, 90, 180, 270) perfect rotations.
There is still a little "burpiness" on some angle rotations, but it's good enough for what i'm planning on using it for.